+++ src/games/planetwars/visualizer/.jsbeautifyrc
@@ -1,0 +1,10 @@
+{
+    "indent_size": 2,
+    "indent_char": " ",
+    "css": {
+      "indent_size": 4
+    },
+    "html": {
+      "indent_size": 2
+    }
+}
