+++ src/games/planetwars/visualizer/index.html
@@ -3,116 +3,12 @@
 
 <head>
   <meta charset="utf-8">
-  <title>Visualization MOZAIC</title>
-  <link rel="stylesheet" href="style.css">
+  <title>Planet Wars</title>
 </head>
 
 <body>
-  <div id="visualizer">
-
-    <!-- -->
-    <svg id="score"></svg>
-
-    <!-- Hide/unhide button for scoreboard -->
-    <button title="Hide scoreboard" type="button" name="button" id="hide_score" class="control control-button close">
-      <i class="fa fa-times" aria-hidden="true"></i>
-    </button>
-
-    <button title="Show scoreboard" type="button" name="button" id="unhide_score" class="control control-button">
-      <i class="fa fa-chevron-left" aria-hidden="true"></i>
-    </button>
-
-    <!-- Hide/unhide button -->
-    <button title="Hide controls" type="button" name="button" id="hide" class="control control-button">
-      <i class="fa fa-chevron-down" aria-hidden="true"></i>
-    </button>
-
-    <button title="Show controls" type="button" name="button" id="unhide" class="control control-button">
-      <i class="fa fa-chevron-up" aria-hidden="true"></i>
-    </button>
-
-    <div id="controlbar">
-
-      <input type="range" id="turn_slider" value="0" class="control" />
-
-      <div id="control-buttons">
-        <div id="turncontrols">
-
-          <!-- Rewind button -->
-          <button type="button" name="button" id="tostart" class="control control-button">
-            <i class="fa fa-fast-backward" aria-hidden="true"></i>
-          </button>
-
-          <!-- Previous turn button -->
-          <button type="button" name="button" id="previous" class="control control-button">
-            <i class="fa fa-step-backward" aria-hidden="true"></i>
-          </button>
-
-          <!-- Play -->
-          <button type="button" name="button" id="play" class="control control-button">
-            <i class="fa fa-play" aria-hidden="true"></i>
-          </button>
-
-          <!-- Pause button -->
-          <button type="button" name="button" id="pause" class="control control-button">
-            <i class="fa fa-pause" aria-hidden="true"></i>
-          </button>
-
-          <!-- Next turn button -->
-          <button type="button" name="button" id="next" class="control control-button">
-            <i class="fa fa-step-forward" aria-hidden="true"></i>
-          </button>
-
-          <!-- Forward button -->
-          <button type="button" name="button" id="toend" class="control control-button">
-            <i class="fa fa-fast-forward" aria-hidden="true"></i>
-          </button>
-          <p id="turn_progress">100 / 100</p>
-        </div>
-
-        <!-- Select game or map -->
-        <div id="filecontrols">
-          <div class="file-select">
-            <span>Select game</span>
-            <input id="game-select" type="file" class="control file-finder" />
-          </div>
-          <div class="file-select">
-            <span>Select map</span>
-            <input id="map-select" type="file" class="control file-finder" />
-          </div>
-        </div>
-
-        <div id="speedcontrols">
-          <p class="speed">Speed x1</p>
-          <!-- Speed down button -->
-          <button type="button" name="button" id="speeddown" class="control control-button">
-            <i class="fa fa-minus" aria-hidden="true"></i>
-          </button>
-          <!-- Speedup button -->
-          <button type="button" name="button" id="speedup" class="control control-button">
-            <i class="fa fa-plus" aria-hidden="true"></i>
-          </button>
-        </div>
-      </div>
-    </div>
-
-    <!-- The SVG where the game gets rendered" -->
-    <svg id="game" font-family="sans-serif" font-size="10" text-anchor="middle">
-    </svg>
-
-    <div id="end_card">
-      <button title="Hide end card" type="button" name="button" id="hide_card" class="control control-button close">
-        <i class="fa fa-times" aria-hidden="true"></i>
-      </button>
-      <p>
-        Game over:
-        <br/>
-        <span class="winner">winner</span> wins!
-      </p>
-    </div>
-  </div>
-
-  <script src="./bundle.js"></script>
+  <div id="visualizer"></div>
+  <script src="./app/bundle.js"></script>
 </body>
 
 </html>
