+++ src/games/planetwars/visualizer/src/app.js
@@ -1,6 +1,7 @@
+const ReactDOM = require('react-dom');
+const h = require('react-hyperscript');
 const Visualizer = require('./visualizer');
 
 window.onload = function() {
-  new Visualizer();
+  ReactDOM.render(h(Visualizer), document.getElementById('visualizer'));
 };
-
