+++ src/games/planetwars/visualizer/src/util/spacemath.js
@@ -41,27 +41,6 @@ class SpaceMath {
   }
 }
 
-class DataBinder {
-  constructor(initial) {
-    this.value = initial;
-    this.callbacks = [];
-  }
-
-  registerCallback(callback) {
-    this.callbacks.push(callback);
-  }
-
-  update(value) {
-    this.value = value;
-    this._propagate();
-  }
-
-  _propagate() {
-    this.callbacks.forEach(c => c(this.value));
-  }
-}
-
 module.exports = {
-  DataBinder,
   SpaceMath
-};
+}
